demo.py: rebuild around the machinery the notebook uses; fix feature-label loader
Demo: 5 steps (load -> replacement model -> graph -> propagate -> constrained
sweep) instead of a 10-step tour. Drops the progressive curve (unused by the
notebook) and the re-prune showcase; adds the propagate-vs-constrained contrast
the old script never made. Feature selection is now semantic (the features that
write the answer, below a depth cap) — influence rank picked L33-35 answer-writers,
which left sweep_patch_end_layer exactly one end layer, so the sweep said nothing.
Eager decoders, seeded, --lazy-decoder escape hatch.

feature_labels: a zero-length blob means the repo stores no label for that feature
(0.4% of them), not that the data is corrupt; skip it at debug level instead of
warning. load_feature_examples caught only (IndexError, ValueError), so this would
have crashed it — guarded by a test. attribution_graph: detach the influence-seed
probs. hf_loader: torch_dtype -> dtype (transformers >=4.56).

Tests: examples API-conformance guard (21 call sites) + feature-label blob tests.

+++ src/llm_circuits/models/hf_loader.py
@@ -74,7 +74,7 @@ def load_model_and_tokenizer(
     )
     model = AutoModelForCausalLM.from_pretrained(
         model_id,
-        torch_dtype=dtype,
+        dtype=dtype,  # `torch_dtype` is deprecated since transformers 4.56 (see pyproject floor)
         device_map=device_map,
         trust_remote_code=trust_remote_code,
         cache_dir=resolved_cache_dir,
